refactor : moveCar 로직은 CarTest.kt 에서 별도로 확인하도록 변경
