+++ src/test/kotlin/controller/RaceControllerTest.kt
@@ -3,14 +3,12 @@ package controller
 import model.Car
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import view.InOutView
 import java.io.ByteArrayOutputStream
 import java.io.OutputStream
 import java.io.PrintStream
-import java.util.Random
 
 class RaceControllerTest {
     private lateinit var outputStream: OutputStream
@@ -34,14 +32,6 @@ class RaceControllerTest {
     val cars: MutableList<Car> = mutableListOf(Car("hwan", 2), Car("sia", 4))
     val inOutController = InOutController(InOutView())
     val raceController = RaceController(3, cars, inOutController)
-
-    @Test
-    fun `랜덤 숫자가 4 이상일 때 전진한다`() {
-        val random = Random(1234)
-        val position = cars[0].currentPosition
-        cars[0].moveCar(random)
-        assertEquals(position + 1, cars[0].currentPosition)
-    }
 
     @Test
     fun `가장 많이 움직인 자동차가 우승한다`() {
